test(reflection): relocate inverse-detection tests to inverse-associations
Move 8 TS-only inverse-detection tests out of reflection.test.ts, where they
matched no test in reflection_test.rb (residual RFC 0043 `extra`). The 3
name-matching tests already exist canonically in inverse-associations.test.ts,
so they are dropped rather than duplicated. The other 5 exercise the same
automatic-inverse machinery with no verbatim Rails analog and are relocated
into inverse-associations.test.ts as deliberate extras (bespoke model names
registered under unique keys to avoid clobbering the canonical models shared
across the describe).

reflection_test.rb extra drops by 8; reflection matched and
inverse_associations_test.rb matched are unchanged.
