+++ packages/activerecord/src/reflection.test.ts
@@ -1017,267 +1017,6 @@ describe("ReflectionTest", () => {
     // UNPORTED: relies on Ruby const_missing mechanism — no JS equivalent.
   });
 
-  it("has one and belongs to should find inverse automatically", () => {
-    class Car extends Base {
-      declare bulb: Bulb | null;
-      declare loadHasOne: (name: "bulb") => Promise<Bulb | null>;
-
-      static {
-        this.attribute("id", "integer");
-        this.hasOne("bulb", {});
-      }
-    }
-    class Bulb extends Base {
-      declare car_id: number | null;
-      declare car: Car | null;
-      declare loadBelongsTo: (name: "car") => Promise<Car | null>;
-
-      static {
-        this.attribute("id", "integer");
-        this.attribute("car_id", "integer");
-        this.belongsTo("car", {});
-      }
-    }
-    registerModel(Car);
-    registerModel(Bulb);
-
-    const carRef = reflectOnAssociation(Car, "bulb")!;
-    const bulbRef = reflectOnAssociation(Bulb, "car")!;
-
-    expect(carRef.hasInverse()).toBe(true);
-    expect(carRef.inverseOf()!.name).toBe("car");
-
-    expect(bulbRef.hasInverse()).toBe(true);
-    expect(bulbRef.inverseOf()!.name).toBe("bulb");
-  });
-
-  it("has many and belongs to should find inverse automatically", () => {
-    class Comment extends Base {
-      static {
-        this.attribute("id", "integer");
-      }
-    }
-    class Rating extends Base {
-      declare comment_id: number | null;
-      declare comment: Comment | null;
-      declare loadBelongsTo: (name: "comment") => Promise<Comment | null>;
-
-      static {
-        this.attribute("id", "integer");
-        this.attribute("comment_id", "integer");
-        this.belongsTo("comment", {});
-      }
-    }
-    registerModel(Comment);
-    registerModel(Rating);
-    Associations.hasMany.call(Comment, "ratings", {});
-
-    const commentRef = reflectOnAssociation(Comment, "ratings")!;
-    expect(commentRef.hasInverse()).toBe(true);
-    expect(commentRef.inverseOf()!.name).toBe("comment");
-  });
-
-  it("has one and belongs to with non default foreign key should not find inverse automatically", () => {
-    class User extends Base {
-      static {
-        this.attribute("id", "integer");
-      }
-    }
-    class Room extends Base {
-      declare owner_id: number | null;
-      declare owner: User | null;
-      declare loadBelongsTo: (name: "owner") => Promise<User | null>;
-
-      static {
-        this.attribute("id", "integer");
-        this.attribute("owner_id", "integer");
-        this.belongsTo("owner", { className: "User", foreignKey: "owner_id" });
-      }
-    }
-    registerModel(User);
-    registerModel(Room);
-    Associations.hasOne.call(User, "ownedRoom", { foreignKey: "owner_id" });
-
-    const ownerRef = reflectOnAssociation(Room, "owner")!;
-    expect(ownerRef.hasInverse()).toBe(false);
-  });
-
-  it("through association should not find inverse automatically", () => {
-    class Doctor extends Base {
-      declare appointments: AssociationProxy<Appointment>;
-      declare patients: AssociationProxy<Base>;
-
-      static {
-        this.attribute("id", "integer");
-        this.hasMany("appointments", {});
-        this.hasMany("patients", { through: "appointments" });
-      }
-    }
-    class Appointment extends Base {
-      declare doctor_id: number | null;
-      declare patient_id: number | null;
-      declare doctor: Doctor | null;
-      declare patient: Patient | null;
-      declare loadBelongsTo: ((name: "doctor") => Promise<Doctor | null>) &
-        ((name: "patient") => Promise<Patient | null>);
-
-      static {
-        this.attribute("id", "integer");
-        this.attribute("doctor_id", "integer");
-        this.attribute("patient_id", "integer");
-        this.belongsTo("doctor", {});
-        this.belongsTo("patient", {});
-      }
-    }
-    class Patient extends Base {
-      static {
-        this.attribute("id", "integer");
-      }
-    }
-    registerModel(Doctor);
-    registerModel(Appointment);
-    registerModel(Patient);
-
-    const patientsRef = reflectOnAssociation(Doctor, "patients")!;
-    expect(patientsRef.hasInverse()).toBe(false);
-  });
-
-  it("polymorphic belongs to should not find inverse automatically", () => {
-    class Tag extends Base {
-      declare taggable_id: number | null;
-      declare taggable_type: string | null;
-
-      static {
-        this.attribute("id", "integer");
-        this.attribute("taggable_id", "integer");
-        this.attribute("taggable_type", "string");
-      }
-    }
-    class Post extends Base {
-      static {
-        this.attribute("id", "integer");
-      }
-    }
-    registerModel(Tag);
-    registerModel(Post);
-    Associations.belongsTo.call(Tag, "taggable", { polymorphic: true });
-    Associations.hasMany.call(Post, "tags", { as: "taggable" });
-
-    const taggableRef = reflectOnAssociation(Tag, "taggable")!;
-    expect(taggableRef.hasInverse()).toBe(false);
-  });
-
-  it("explicit inverse of false disables automatic detection", () => {
-    class Parent extends Base {
-      static {
-        this.attribute("id", "integer");
-      }
-    }
-    class Child extends Base {
-      declare parent_id: number | null;
-
-      static {
-        this.attribute("id", "integer");
-        this.attribute("parent_id", "integer");
-      }
-    }
-    registerModel(Parent);
-    registerModel(Child);
-    Associations.hasMany.call(Parent, "children", { className: "Child", inverseOf: false });
-    Associations.belongsTo.call(Child, "parent", {});
-
-    const childrenRef = reflectOnAssociation(Parent, "children")!;
-    expect(childrenRef.hasInverse()).toBe(false);
-  });
-
-  it("has many with scope should not find inverse automatically unless automatic scope inversing", () => {
-    // Without automatic_scope_inversing, scoped associations should not find inverse
-    {
-      class Company extends Base {
-        static {
-          this.attribute("id", "integer");
-        }
-      }
-      class Contract extends Base {
-        declare company_id: number | null;
-        declare company: Company | null;
-        declare loadBelongsTo: (name: "company") => Promise<Company | null>;
-
-        static {
-          this.attribute("id", "integer");
-          this.attribute("company_id", "integer");
-          this.belongsTo("company", {});
-        }
-      }
-      registerModel(Company);
-      registerModel(Contract);
-      const scopeFn = (rel: any) => rel;
-      Associations.hasMany.call(Company, "contracts", { scope: scopeFn });
-
-      const contractsRef = reflectOnAssociation(Company, "contracts")!;
-      expect(contractsRef.hasInverse()).toBe(false);
-    }
-
-    // With automatic_scope_inversing enabled, scoped associations should find inverse
-    {
-      class Company2 extends Base {
-        static {
-          this.attribute("id", "integer");
-        }
-      }
-      class Contract2 extends Base {
-        declare company2_id: number | null;
-        declare company2: Company2 | null;
-        declare loadBelongsTo: (name: "company2") => Promise<Company2 | null>;
-
-        static automaticScopeInversing = true;
-        static {
-          this.attribute("id", "integer");
-          this.attribute("company2_id", "integer");
-          this.belongsTo("company2", { className: "Company2" });
-        }
-      }
-      registerModel("Company2", Company2);
-      registerModel("Contract2", Contract2);
-      const scopeFn = (rel: any) => rel;
-      Associations.hasMany.call(Company2, "contract2s", { scope: scopeFn, className: "Contract2" });
-
-      const contractsRef = reflectOnAssociation(Company2, "contract2s")!;
-      expect(contractsRef.hasInverse()).toBe(true);
-      expect(contractsRef.inverseOf()!.name).toBe("company2");
-    }
-  });
-
-  it("scoped belongs to on inverse side blocks automatic inverse", () => {
-    // Scopes on the inverse (belongs_to) side always block automatic detection,
-    // even when automatic_scope_inversing is enabled
-    class Publisher extends Base {
-      declare magazines: AssociationProxy<Magazine>;
-
-      static automaticScopeInversing = true;
-      static {
-        this.attribute("id", "integer");
-        this.hasMany("magazines", {});
-      }
-    }
-    class Magazine extends Base {
-      declare publisher_id: number | null;
-
-      static automaticScopeInversing = true;
-      static {
-        this.attribute("id", "integer");
-        this.attribute("publisher_id", "integer");
-      }
-    }
-    registerModel(Publisher);
-    registerModel(Magazine);
-    const scopeFn = (rel: any) => rel;
-    Associations.belongsTo.call(Magazine, "publisher", { scope: scopeFn });
-
-    const magazinesRef = reflectOnAssociation(Publisher, "magazines")!;
-    expect(magazinesRef.hasInverse()).toBe(false);
-  });
-
   it("human name", () => {
     class Post extends Base {
       declare title: string | null;
